requirements.txt
openai==0.28
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
beautifulsoup4